+++ bikeability/operator_worker.py
@@ -4,7 +4,6 @@ from typing import List, Dict, Tuple
 
 import geopandas as gpd
 import shapely
-
 from climatoology.base.operator import ComputationResources, Concern, Info, Operator, PluginAuthor, _Artifact
 from ohsome import OhsomeClient
 from semver import Version
@@ -14,11 +13,11 @@ from bikeability.artifact import (
 )
 from bikeability.input import ComputeInputBikeability
 from bikeability.utils import (
-    construct_filters,
     fetch_osm_data,
     PathCategory,
     apply_path_category_filters,
     boost_route_members,
+    ohsome_filter,
 )
 
 log = logging.getLogger(__name__)
@@ -47,8 +46,8 @@ class OperatorBikeability(Operator[ComputeInputBikeability]):
             ],
             version=str(Version(0, 0, 1)),
             concerns=[Concern.MOBILITY_CYCLING],
-            purpose='This is a dummy for the rework of the plugin.',
-            methodology='The dummy is based on the functionality of walkability.',
+            purpose=Path('resources/info/purpose.md').read_text(),
+            methodology=Path('resources/info/methodology.md').read_text(),
             sources=Path('resources/literature.bib'),
         )
         log.info(f'Return info {info.model_dump()}')
@@ -70,32 +69,12 @@ class OperatorBikeability(Operator[ComputeInputBikeability]):
     ) -> Tuple[gpd.GeoDataFrame, gpd.GeoDataFrame]:
         log.debug('Extracting paths')
 
-        ohsome_filter = str(
-            '(geometry:line or geometry:polygon) and '
-            '(highway=* or route=ferry) and not '
-            '(cycleway=separate or cycleway=separate or cycleway:both=separate or '
-            '(cycleway:right=separate and cycleway:left=separate) or '
-            '(cycleway:right=separate and cycleway:left=no) or (cycleway:right=no and cycleway:left=separate))'
-        )
+        paths_line = fetch_osm_data(aoi, ohsome_filter('line'), self.ohsome)
+        paths_line['category'] = paths_line.apply(apply_path_category_filters, axis=1)
+        paths_line['category'] = boost_route_members(aoi=aoi, paths_line=paths_line, ohsome=self.ohsome)
 
-        osm_data = fetch_osm_data(
-            aoi,
-            ohsome_filter,
-            self.ohsome,
-        )
-
-        osm_data['category'] = osm_data.apply(apply_path_category_filters, axis=1, filters=construct_filters().items())
-
-        paths_line: gpd.GeoDataFrame = osm_data[osm_data.geometry.type.isin(['LineString', 'MultiLineString'])]
-        paths_line = paths_line.reset_index(drop=True)
-        paths_polygon: gpd.GeoDataFrame = osm_data[osm_data.geometry.type.isin(['Polygon', 'MultiPolygon'])]
-        paths_polygon = paths_polygon.reset_index(drop=True)
-
-        paths_line['category'] = boost_route_members(
-            aoi=aoi,
-            paths_line=paths_line,
-            ohsome=self.ohsome,
-        )
+        paths_polygon = fetch_osm_data(aoi, ohsome_filter('polygon'), self.ohsome)
+        paths_polygon['category'] = paths_polygon.apply(apply_path_category_filters, axis=1)
 
         paths_line['rating'] = paths_line.category.apply(lambda category: rating_map[category])
         paths_polygon['rating'] = paths_polygon.category.apply(lambda category: rating_map[category])