+++ test/conftest.py
@@ -60,8 +60,8 @@ def expected_info_output() -> Info:
         ],
         version=Version(0, 0, 1),
         concerns=[Concern.MOBILITY_CYCLING],
-        purpose='This is a dummy for the rework of the plugin.',
-        methodology='The dummy is based on the functionality of walkability.',
+        purpose=Path('resources/info/purpose.md').read_text(),
+        methodology=Path('resources/info/methodology.md').read_text(),
         sources=Path('resources/literature.bib'),
     )
 
@@ -72,14 +72,20 @@ def expected_compute_output(compute_resources) -> List[_Artifact]:
         name='Cycling infrastructure path categories',
         modality=ArtifactModality.MAP_LAYER_GEOJSON,
         file_path=Path(compute_resources.computation_dir / 'cycling_infrastructure_path_categories.geojson'),
-        summary='TBD',
-        description='TBD',
+        summary=Path('resources/info/path_categories/caption.md').read_text(),
+        description=Path('resources/info/path_categories/description.md').read_text(),
         attachments={
             AttachmentType.LEGEND: Legend(
                 legend_data={
-                    'designated': Color('#006837'),
-                    'forbidden': Color('#a50026'),
-                    'not_categorised': Color('grey'),
+                    'designated_exclusive': Color('#313695'),
+                    'designated_shared_with_pedestrians': Color('#5183bb'),
+                    'shared_with_motorised_traffic_walking_speed_(<=15_km/h)': Color('#90c3dd'),
+                    'shared_with_motorised_traffic_low_speed_(<=30_km/h)': Color('#d4edf4'),
+                    'shared_with_motorised_traffic_medium_speed_(<=50_km/h)': Color('#fffebe'),
+                    'shared_with_motorised_traffic_high_speed_(<=100_km/h)': Color('#fed283'),
+                    'requires_dismounting': Color('#f88c51'),
+                    'not_bikeable': Color('#dd3d2d'),
+                    'unknown': Color('grey'),
                 }
             )
         },
@@ -109,15 +115,22 @@ def operator():
 @pytest.fixture
 def ohsome_api(responses_mock):
     with (
-        open('resources/test/ohsome_line_and_polygon_response.geojson', 'r') as line_and_polygon_file,
+        open('resources/test/ohsome_line_response.geojson', 'r') as line_file,
+        open('resources/test/ohsome_polygon_response.geojson', 'r') as polygon_file,
         open('resources/test/ohsome_route_response.geojson', 'r') as route_file,
     ):
-        line_and_polygon_body = line_and_polygon_file.read()
+        line_body = line_file.read()
+        polygon_body = polygon_file.read()
         route_body = route_file.read()
     responses_mock.post(
         'https://api.ohsome.org/v1/elements/geometry',
-        body=line_and_polygon_body,
-        match=[filter_start_matcher('(geometry:line or geometry:polygon)')],
+        body=line_body,
+        match=[filter_start_matcher('geometry:line')],
+    )
+    responses_mock.post(
+        'https://api.ohsome.org/v1/elements/geometry',
+        body=polygon_body,
+        match=[filter_start_matcher('geometry:polygon')],
     )
     responses_mock.post(
         'https://api.ohsome.org/v1/elements/geometry',